Modularize more and add swagger-ui
