+++ scripts/.eslintrc
@@ -1,5 +1,8 @@
 {
     "extends": "../.eslintrc",
+    "env": {
+        "node": true
+    },
     "rules": {
         "no-console": "off"
     }
