+++ tsconfig.eslint.json
@@ -4,5 +4,5 @@
         "rootDir": ".",
         "noEmit": true
     },
-    "include": ["src/**/*", "scripts/*", "static/*"],
+    "include": ["src/**/*", "scripts/*"],
 }
